+++ dev/Knoin/AbstractViews.js
@@ -59,6 +59,7 @@ export class AbstractViewPopup extends AbstractView
 		super('Popups' + name, ViewTypePopup);
 		this.keyScope.scope = name;
 		this.modalVisible = ko.observable(false).extend({ rateLimit: 0 });
+		this.close = () => this.modalVisible(false);
 		addShortcut('escape,close', '', name, () => {
 			if (this.modalVisible() && false !== this.onClose()) {
 				this.close();
@@ -78,8 +79,6 @@ export class AbstractViewPopup extends AbstractView
 	afterShow() {}  // Happens after  showModal() animation transitionend
 	onHide() {}     // Happens before animation transitionend
 	afterHide() {}  // Happens after  animation transitionend
-
-	close() {}
 */
 }