+++ dev/Model/MessageCollection.js
@@ -1,5 +1,6 @@
 import { AbstractCollectionModel } from 'Model/AbstractCollection';
 import { MessageModel } from 'Model/Message';
+import { MessageUserStore } from 'Stores/User/Message';
 
 import {
 	MessageFlagsCache
@@ -31,8 +32,11 @@ export class MessageCollectionModel extends AbstractCollectionModel
 	 * @returns {MessageCollectionModel}
 	 */
 	static reviveFromJson(object, cached) {
+		let msg = MessageUserStore.message();
 		return super.reviveFromJson(object, message => {
-			message = MessageModel.reviveFromJson(message);
+			// If message is currently viewed, use that.
+			// Maybe then use msg.revivePropertiesFromJson(message) ?
+			message = (msg && msg.hash === message.hash) ? msg : MessageModel.reviveFromJson(message);
 			if (message) {
 				message.deleted(false);
 				cached ? MessageFlagsCache.initMessage(message) : MessageFlagsCache.store(message);