+++ dev/View/User/MailBox/MessageList.js
@@ -162,12 +162,12 @@ export class MailMessageList extends AbstractViewRight {
 				let list = [], current, sort = FolderUserStore.sortMode() || 'DATE';
 				if (sort.includes('FROM')) {
 					MessagelistUserStore.forEach(msg => {
-						let email = msg.from?.[0].email;
+						let email = msg.from[0].email;
 						if (!current || email != current.id) {
 							current = {
 								id: email,
 								label: msg.from[0].toLine(),
-								search: 'from=' + msg.from[0].email,
+								search: 'from=' + email,
 								messages: []
 							};
 							list.push(current);
@@ -235,7 +235,8 @@ export class MailMessageList extends AbstractViewRight {
 
 		this.selector.on('ItemSelect', message => {
 			if (message) {
-				populateMessageBody(MessageModel.fromMessageListItem(message));
+//				populateMessageBody(message.clone());
+				populateMessageBody(message);
 			} else {
 				MessageUserStore.message(null);
 			}