+++ snappymail/v/0.0.0/app/libraries/MailSo/Base/Http.php
@@ -185,7 +185,7 @@ class Http
 			static::$bCache = true;
 			\header('Expires: Mon, 26 Jul 1997 05:00:00 GMT');
 			\header('Last-Modified: '.\gmdate('D, d M Y H:i:s').' GMT');
-			\header('Cache-Control: no-store, no-cache, must-revalidate, max-age=0, post-check=0, pre-check=0');
+			\header('Cache-Control: no-store');
 			\header('Pragma: no-cache');
 		}
 	}