+++ snappymail/v/0.0.0/app/libraries/MailSo/Imap/SearchCriterias.php
@@ -316,8 +316,12 @@ abstract class SearchCriterias
 
 	public static function escapeSearchString(\MailSo\Imap\ImapClient $oImapClient, string $sSearch) : string
 	{
-		return !\MailSo\Base\Utils::IsAscii($sSearch)
-			? '{'.\strlen($sSearch).'}'."\r\n".$sSearch : $oImapClient->EscapeString($sSearch);
+		// https://github.com/the-djmaze/snappymail/issues/836
+//		return $oImapClient->EscapeString($sSearch);
+//		return \MailSo\Base\Utils::IsAscii($sSearch) || $oImapClient->hasCapability('QQMail'))
+		return (\MailSo\Base\Utils::IsAscii($sSearch) || !$oImapClient->hasCapability('LITERAL+'))
+			? $oImapClient->EscapeString($sSearch)
+			: '{'.\strlen($sSearch).'}'."\r\n{$sSearch}";
 	}
 
 	private static function parseSearchDate(string $sDate) : int