+++ snappymail/v/0.0.0/app/libraries/RainLoop/Actions/User.php
@@ -181,6 +181,7 @@ trait User
 		$this->setSettingsFromParams($oSettings, 'ViewImages', 'string');
 		$this->setSettingsFromParams($oSettings, 'ViewImagesWhitelist', 'string');
 		$this->setSettingsFromParams($oSettings, 'RemoveColors', 'bool');
+		$this->setSettingsFromParams($oSettings, 'AllowStyles', 'bool');
 		$this->setSettingsFromParams($oSettings, 'ListInlineAttachments', 'bool');
 		$this->setSettingsFromParams($oSettings, 'CollapseBlockquotes', 'bool');
 		$this->setSettingsFromParams($oSettings, 'MaxBlockquotesLevel', 'int');